@@ -5,7 +5,7 @@ version: 1.1.1+3
 publish_to: 'none'
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
   flutter: 3.0.0
 
 dependencies:
@@ -45,6 +45,7 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.5
+  path: ^1.8.2
 
 dev_dependencies:
   flutter_test:
