@@ -2,7 +2,7 @@ class Channel {
 
 String name;
 String avatarImage;
-bool isCheck;
+bool? isCheck;
 
 Channel(this.name, this.avatarImage, this.isCheck);
 }
