@@ -1,12 +1,12 @@
 class IndexingInfo {
   var entries; //Todo: determine what entries are & choose apropriate type
-  String time;
-  bool error;
-  bool done;
-  double parserProgress;
-  double indexerProgress;
-  int parsingProgress;
-  int indexingProgress;
+  String? time;
+  bool? error;
+  bool? done;
+  double? parserProgress;
+  double? indexerProgress;
+  int? parsingProgress;
+  int? indexingProgress;
 
   IndexingInfo();
 
