@@ -2,6 +2,6 @@ import 'package:flutter_ws/model/query_info.dart';
 import 'package:flutter_ws/model/video.dart';
 
 class QueryResult {
-  Iterable<Video> videos;
-  QueryInfo queryInfo;
+  late Iterable<Video> videos;
+  late QueryInfo queryInfo;
 }
