@@ -15,13 +15,13 @@ import 'package:video_player/video_player.dart';
 import 'TVPlayerController.dart';
 
 class FlutterVideoPlayer extends StatefulWidget {
-  String videoId;
+  String? videoId;
   Video video;
-  VideoEntity videoEntity;
-  CustomChewieController chewieController;
-  DatabaseManager databaseManager;
-  VideoProgressEntity progressEntity;
-  AppSharedState appSharedState;
+  VideoEntity? videoEntity;
+  late CustomChewieController chewieController;
+  DatabaseManager? databaseManager;
+  VideoProgressEntity? progressEntity;
+  late AppSharedState appSharedState;
   bool isAlreadyPlayingDifferentVideoOnTV = false;
 
   final Logger log = new Logger('FlutterVideoPlayer');
@@ -29,15 +29,15 @@ class FlutterVideoPlayer extends StatefulWidget {
   final Logger logger = new Logger('FlutterVideoPlayer');
 
   FlutterVideoPlayer(BuildContext context, AppSharedState appSharedState,
-      this.video, VideoEntity entity, VideoProgressEntity progress) {
-    this.videoId = video != null ? video.id : entity.id;
-    this.databaseManager = appSharedState.appState.databaseManager;
+      this.video, VideoEntity? entity, VideoProgressEntity? progress) {
+    this.videoId = video != null ? video.id : entity!.id;
+    this.databaseManager = appSharedState.appState!.databaseManager;
     this.progressEntity = progress;
     this.videoEntity = entity;
     this.appSharedState = appSharedState;
 
-    if (appSharedState.appState.isCurrentlyPlayingOnTV &&
-        videoId != appSharedState.appState.tvCurrentlyPlayingVideo.id) {
+    if (appSharedState.appState!.isCurrentlyPlayingOnTV &&
+        videoId != appSharedState.appState!.tvCurrentlyPlayingVideo.id) {
       isAlreadyPlayingDifferentVideoOnTV = true;
     }
   }
@@ -47,12 +47,12 @@ class FlutterVideoPlayer extends StatefulWidget {
 }
 
 class _FlutterVideoPlayerState extends State<FlutterVideoPlayer> {
-  String videoUrl;
+  String? videoUrl;
   // castNewVideoToTV indicates that the currently playing video on the TV
   // should be replaced
   bool castNewVideoToTV = false;
-  static VideoPlayerController videoController;
-  static TvPlayerController tvVideoController;
+  static VideoPlayerController? videoController;
+  static TvPlayerController? tvVideoController;
 
   @override
   Widget build(BuildContext context) {
@@ -79,15 +79,15 @@ class _FlutterVideoPlayerState extends State<FlutterVideoPlayer> {
     super.dispose();
   }
 
-  String getVideoUrl(Video video, VideoEntity entity) {
+  String? getVideoUrl(Video video, VideoEntity? entity) {
     if (video != null) {
-      if (video.url_video_hd != null && video.url_video_hd.isNotEmpty) {
+      if (video.url_video_hd != null && video.url_video_hd!.isNotEmpty) {
         return video.url_video_hd;
       } else {
         return video.url_video;
       }
     } else {
-      if (entity.url_video_hd != null && entity.url_video_hd.isNotEmpty) {
+      if (entity!.url_video_hd != null && entity.url_video_hd!.isNotEmpty) {
         return entity.url_video_hd;
       } else {
         return entity.url_video;
@@ -97,49 +97,49 @@ class _FlutterVideoPlayerState extends State<FlutterVideoPlayer> {
 
   void initTvVideoController() {
     tvVideoController = new TvPlayerController(
-      widget.appSharedState.appState.availableTvs,
-      widget.appSharedState.appState.samsungTVCastManager,
-      widget.appSharedState.appState.databaseManager,
+      widget.appSharedState.appState!.availableTvs,
+      widget.appSharedState.appState!.samsungTVCastManager,
+      widget.appSharedState.appState!.databaseManager,
       videoUrl,
       widget.video != null
           ? widget.video
-          : Video.fromMap(widget.videoEntity.toMap()),
+          : Video.fromMap(widget.videoEntity!.toMap()),
       widget.progressEntity != null
-          ? new Duration(milliseconds: widget.progressEntity.progress)
+          ? new Duration(milliseconds: widget.progressEntity!.progress!)
           : new Duration(milliseconds: 0),
     );
 
-    if (widget.appSharedState.appState.targetPlatform ==
+    if (widget.appSharedState.appState!.targetPlatform ==
         TargetPlatform.android) {
-      tvVideoController.startTvDiscovery();
+      tvVideoController!.startTvDiscovery();
     }
 
     // replace the currently playing video on TV
-    if (widget.appSharedState.appState.isCurrentlyPlayingOnTV &&
+    if (widget.appSharedState.appState!.isCurrentlyPlayingOnTV &&
         castNewVideoToTV) {
-      widget.appSharedState.appState.samsungTVCastManager.stop();
-      tvVideoController.initialize();
-      tvVideoController.startPlayingOnTV();
+      widget.appSharedState.appState!.samsungTVCastManager.stop();
+      tvVideoController!.initialize();
+      tvVideoController!.startPlayingOnTV();
       return;
     }
 
     // case: do not replace the currently playing video on TV
-    if (widget.appSharedState.appState.isCurrentlyPlayingOnTV) {
-      tvVideoController.initialize();
+    if (widget.appSharedState.appState!.isCurrentlyPlayingOnTV) {
+      tvVideoController!.initialize();
     }
   }
 
   void initVideoPlayerController(BuildContext context) {
     if (videoController != null) {
-      videoController.dispose();
+      videoController!.dispose();
     }
     // always use network datasource if should be casted to TV
     // TV needs accessible video URL
     if (widget.videoEntity == null ||
-        widget.appSharedState.appState.isCurrentlyPlayingOnTV &&
+        widget.appSharedState.appState!.isCurrentlyPlayingOnTV &&
             widget.video != null) {
       videoController = VideoPlayerController.network(
-        videoUrl,
+        videoUrl!,
       );
 
       Map<String, Object> event = {"key": "PLAY_VIDEO_NETWORK", "count": 1};
@@ -149,14 +149,14 @@ class _FlutterVideoPlayerState extends State<FlutterVideoPlayer> {
     }
 
     String path;
-    if (widget.appSharedState.appState.targetPlatform ==
+    if (widget.appSharedState.appState!.targetPlatform ==
         TargetPlatform.android) {
-      path = widget.videoEntity.filePath + "/" + widget.videoEntity.fileName;
+      path = widget.videoEntity!.filePath! + "/" + widget.videoEntity!.fileName!;
     } else {
-      path = widget.appSharedState.appState.localDirectory.path +
+      path = widget.appSharedState.appState!.localDirectory!.path +
           "/MediathekView" +
           "/" +
-          widget.videoEntity.fileName;
+          widget.videoEntity!.fileName!;
     }
 
     Uri videoUri = new Uri.file(path);
@@ -169,7 +169,7 @@ class _FlutterVideoPlayerState extends State<FlutterVideoPlayer> {
               "Cannot play video from file. File does not exist: " +
                   file.uri.toString());
           videoController = VideoPlayerController.network(
-            videoUrl,
+            videoUrl!,
           );
         }
       },
@@ -184,20 +184,20 @@ class _FlutterVideoPlayerState extends State<FlutterVideoPlayer> {
   void initChewieController() {
     widget.chewieController = new CustomChewieController(
         context: context,
-        videoPlayerController: videoController,
+        videoPlayerController: videoController!,
         tvPlayerController: tvVideoController,
         looping: false,
-        startAt: tvVideoController.startAt,
+        startAt: tvVideoController!.startAt,
         customControls: new CustomVideoControls(
             backgroundColor: Color.fromRGBO(41, 41, 41, 0.7),
             iconColor: Color(0xffffbf00)),
         fullScreenByDefault: false,
         allowedScreenSleep: false,
         isCurrentlyPlayingOnTV:
-            widget.appSharedState.appState.isCurrentlyPlayingOnTV,
+            widget.appSharedState.appState!.isCurrentlyPlayingOnTV,
         video: widget.video != null
             ? widget.video
-            : Video.fromMap(widget.videoEntity.toMap()));
+            : Video.fromMap(widget.videoEntity!.toMap()));
   }
 
   AlertDialog _showDialog(BuildContext context) {
@@ -215,16 +215,16 @@ class _FlutterVideoPlayerState extends State<FlutterVideoPlayer> {
             // replace widget.video with the currently playing video
             // to not interrupt the video playback
             widget.video =
-                widget.appSharedState.appState.tvCurrentlyPlayingVideo;
+                widget.appSharedState.appState!.tvCurrentlyPlayingVideo;
 
             // get the video entity
             widget.videoEntity = await widget
-                .appSharedState.appState.databaseManager
+                .appSharedState.appState!.databaseManager
                 .getDownloadedVideo(widget.videoId);
 
             // get the video progress
             widget.progressEntity = await widget
-                .appSharedState.appState.databaseManager
+                .appSharedState.appState!.databaseManager
                 .getVideoProgressEntity(widget.video.id);
 
             // start initializing players with the video playing on the TV
@@ -234,7 +234,7 @@ class _FlutterVideoPlayerState extends State<FlutterVideoPlayer> {
         ElevatedButton(
           child: const Text('Ja'),
           onPressed: () {
-            widget.appSharedState.appState.samsungTVCastManager.stop();
+            widget.appSharedState.appState!.samsungTVCastManager.stop();
 
             setState(() {
               widget.isAlreadyPlayingDifferentVideoOnTV = false;
