@@ -12,10 +12,10 @@ class LoadingListPage extends StatelessWidget {
   Widget build(BuildContext context) {
     Widget loadingCard = getLoadingCard();
     int num = determineNumberOfNeededTilesToFillScreen(context, 130);
-    List<int> children = new List(num);
+    List<int?> children = List.filled(num, 0);
 
-    return new SingleChildScrollView(
-      child: new Column(
+    return SingleChildScrollView(
+      child: Column(
         mainAxisSize: MainAxisSize.max,
         crossAxisAlignment: CrossAxisAlignment.start,
         children: children
@@ -38,71 +38,71 @@ class LoadingListPage extends StatelessWidget {
     final card = getListCard(cardContent);
     Widget dummyChannelThumbnail = getDummyChannelThumbnail();
 
-    return new Container(
+    return Container(
       margin: const EdgeInsets.symmetric(
         vertical: 8.0,
         horizontal: 8.0,
       ),
-      child: new Stack(
+      child: Stack(
         children: <Widget>[card, dummyChannelThumbnail],
       ),
     );
   }
 
   Container getCardContent() {
-    return new Container(
-      margin: new EdgeInsets.only(top: 12.0, bottom: 12.0),
-      child: new Column(
+    return Container(
+      margin: EdgeInsets.only(top: 12.0, bottom: 12.0),
+      child: Column(
         mainAxisSize: MainAxisSize.min,
         crossAxisAlignment: CrossAxisAlignment.start,
         children: <Widget>[
-          new Container(height: 4.0),
-          new Flexible(
-            child: new Container(
-              margin: new EdgeInsets.only(left: 40.0, right: 9.0),
-              child: new Container(
+          Container(height: 4.0),
+          Flexible(
+            child: Container(
+              margin: EdgeInsets.only(left: 40.0, right: 9.0),
+              child: Container(
                 color: Colors.grey,
                 constraints: BoxConstraints.expand(width: 50, height: 10),
               ),
             ),
           ),
-          new Container(height: 20.0),
-          new Flexible(
-            child: new Container(
-              margin: new EdgeInsets.only(left: 40.0, right: 40.0),
-              child: new Container(
+          Container(height: 20.0),
+          Flexible(
+            child: Container(
+              margin: EdgeInsets.only(left: 40.0, right: 40.0),
+              child: Container(
                 color: Colors.grey,
                 constraints: BoxConstraints.expand(height: 11),
               ),
             ),
           ),
-          new Container(
-              padding: new EdgeInsets.only(left: 40.0, right: 12.0),
+          Container(
+              padding: EdgeInsets.only(left: 40.0, right: 12.0),
               height: 10),
-          new Flexible(
-              child: new Container(
+          Flexible(
+              child: Container(
                   margin:
-                      new EdgeInsets.symmetric(vertical: 8.0, horizontal: 40.0),
+                      EdgeInsets.symmetric(vertical: 8.0, horizontal: 40.0),
                   height: 2.0,
                   color: Colors.grey)),
-          new Container(height: 4.0),
-          new Container(
+          Container(height: 4.0),
+          Container(
             constraints: BoxConstraints.loose(Size.fromHeight(20.0)),
-            padding: new EdgeInsets.only(left: 40.0, right: 5.0),
-            child: new Row(
+            padding: EdgeInsets.only(left: 40.0, right: 5.0),
+            child: Row(
               children: <Widget>[
-                new Expanded(
+                Expanded(
                     child: _getMetadataRow(
                   width: 20,
-                  icon: new Icon(
+                  icon: Icon(
                     Icons.spa,
                     color: Colors.grey,
                   ),
                 )),
-                new Expanded(
+                Expanded(
                     child: _getMetadataRow(
                   width: 40,
-                  icon: new Icon(
+                  icon: Icon(
                     Icons.access_time,
                     color: Colors.grey,
                   ),
@@ -115,19 +115,19 @@ class LoadingListPage extends StatelessWidget {
     );
   }
 
-  Widget _getMetadataRow({double width, Icon icon}) {
-    return new Row(children: <Widget>[
+  Widget _getMetadataRow({double? width, required Icon icon}) {
+    return Row(children: <Widget>[
       icon,
-      new Container(width: 8.0),
-      new Container(width: width, height: 6, color: Colors.grey[300]),
+      Container(width: 8.0),
+      Container(width: width, height: 6, color: Colors.grey[300]),
     ]);
   }
 
   Container getListCard(Container cardContent) {
-    return new Container(
+    return Container(
       child: cardContent,
-      margin: new EdgeInsets.only(left: 20.0),
-      decoration: new BoxDecoration(
+      margin: EdgeInsets.only(left: 20.0),
+      decoration: BoxDecoration(
         color: Colors.white,
         shape: BoxShape.rectangle,
       ),
@@ -135,12 +135,12 @@ class LoadingListPage extends StatelessWidget {
   }
 
   Container getDummyChannelThumbnail() {
-    return new Container(
-      margin: new EdgeInsets.only(left: 2.0, top: 5.0),
+    return Container(
+      margin: EdgeInsets.only(left: 2.0, top: 5.0),
       alignment: FractionalOffset.topLeft,
       width: 50.0,
       height: 50.0,
-      decoration: new BoxDecoration(
+      decoration: BoxDecoration(
         shape: BoxShape.circle,
         color: Colors.grey,
       ),
